+++ composer.json
@@ -18,19 +18,19 @@
         "ext-mbstring": "*",
         "behat/gherkin": "^4.6.0",
         "behat/transliterator": "^1.2",
-        "symfony/console": "~2.7.40||^2.8.33||~3.3.15||^3.4.3||^4.0.3||^5.0.0",
-        "symfony/config": "~2.3||~3.0||~4.0||~5.0",
-        "symfony/dependency-injection": "~2.1||~3.0||~4.0||~5.0",
-        "symfony/event-dispatcher": "~2.1||~3.0||~4.0||~5.0",
-        "symfony/translation": "~2.3||~3.0||~4.0||~5.0",
-        "symfony/yaml": "~2.1||~3.0||~4.0||~5.0",
+        "symfony/console": "^2.7.51 || ^2.8.33 || ^3.3.15 || ^3.4.3 || ^4.0.3 || ^5.0",
+        "symfony/config": "^2.7.51 || ^3.0 || ^4.0 || ^5.0",
+        "symfony/dependency-injection": "^2.7.51 || ^3.0 || ^4.0 || ^5.0",
+        "symfony/event-dispatcher": "^2.7.51 || ^3.0 || ^4.0 || ^5.0",
+        "symfony/translation": "^2.7.51 || ^3.0 || ^4.0 || ^5.0",
+        "symfony/yaml": "^2.7.51 || ^3.0 || ^4.0 || ^5.0",
         "psr/container": "^1.0",
         "container-interop/container-interop": "^1.2"
     },
 
     "require-dev": {
-        "symfony/process": "~2.5|~3.0|~4.0|~5.0",
-        "phpunit/phpunit": "^4.8.36|^6.3",
+        "symfony/process": "~2.5 || ^3.0 || ^4.0 || ^5.0",
+        "phpunit/phpunit": "^4.8.36 || ^6.3",
         "herrera-io/box": "~1.6.1"
     },
