+++ src/Behat/Behat/Definition/Translator/TranslatorInterface.php
@@ -6,7 +6,7 @@ if (interface_exists(\Symfony\Contracts\Translation\TranslatorInterface::class))
     interface TranslatorInterface extends \Symfony\Contracts\Translation\TranslatorInterface
     {
     }
-} else if (interface_exists(\Symfony\Component\Translation\TranslatorInterface::class)) {
+} elseif (interface_exists(\Symfony\Component\Translation\TranslatorInterface::class)) {
     interface TranslatorInterface extends \Symfony\Component\Translation\TranslatorInterface
     {
     }
