+++ src/Behat/Testwork/EventDispatcher/TestworkEventDispatcher.php
@@ -18,29 +18,16 @@ use Symfony\Component\EventDispatcher\EventDispatcher;
  * @author Konstantin Kudryashov <ever.zet@gmail.com>
  */
 
-if (class_exists(\Symfony\Contracts\EventDispatcher\Event::class)) {
+if (class_exists(\Symfony\Contracts\EventDispatcher\Event::class) && PHP_VERSION_ID > 70200) {
+    // Assert: This is Symfony 5 and PHP >= 7.2
+    include_once __DIR__.'/TestworkEventDispatcherPhp72Trait.php';
 
     final class TestworkEventDispatcher extends EventDispatcher
     {
+        use \TestworkEventDispatcherPhp72Trait;
         const BEFORE_ALL_EVENTS = '*~';
         const AFTER_ALL_EVENTS = '~*';
-
-        /**
-         * {@inheritdoc}
-         */
-        public function dispatch($event, string $eventName = null): object
-        {
-            if (null === $event) {
-                $event = new \Symfony\Contracts\EventDispatcher\Event();
-            }
-            if (method_exists($event, 'setName')) {
-                $event->setName($eventName);
-            }
-
-            $this->callListeners($this->getListeners($eventName), $eventName, $event);
-
-            return $event;
-        }
+        const DISPATCHER_VERSION = 2;
 
         /**
          * {@inheritdoc}
@@ -64,6 +51,7 @@ if (class_exists(\Symfony\Contracts\EventDispatcher\Event::class)) {
     {
         const BEFORE_ALL_EVENTS = '*~';
         const AFTER_ALL_EVENTS = '~*';
+        const DISPATCHER_VERSION = 1;
 
         /**
          * {@inheritdoc}
